test(discoverability): add unit tests for domain discoverability helpers
- Added comprehensive unit tests for bounded_levenshtein and suggestion_for_unknown_domain functions.
- Added tests validating guidance output formatting for missing and unknown domain scenarios.
- Refactored tests to move domain discoverability related tests into a dedicated module.
- Removed inline unit tests from discoverability.rs source file to improve code organization.
- Updated unit test registration in unit.rs for discoverability module.
- Minor fix in actionable_guidance.rs test adjustments to expected next command strings.
